Disabling pygments to avoid warning
The page build completed successfully, but returned the following
warning:

You are attempting to use the 'pygments' highlighter, which is
currently unsupported on GitHub Pages. Your site will use 'rouge' for
highlighting instead. To suppress this warning, change the
'highlighter' value to 'rouge' in your '_config.yml' and ensure the
'pygments' key is unset. For more information, see
https://help.github.com/articles/page-build-failed-config-file-error/#fi
xing-highlighting-errors.

For information on troubleshooting Jekyll see:

 https://help.github.com/articles/troubleshooting-jekyll-builds